requirements.txt
Flask==0.12.2
Flask-API==1.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-MongoEngine==0.9.5
flask-migrate==2.1.1
flask-bcrypt==0.7.1
flask-cors==3.0.3
marshmallow==3.0.0b6
psycopg2==2.7.4
gunicorn==19.7.1
pyjwt==1.5.3
gevent==1.3a2

Flask-Testing==0.7.1
factory-boy==2.10.0
pylint==1.8.2
pylint-quotes==0.1.7
pycodestyle==2.3.1
pydocstyle==2.1.1
pytest==3.5.1
pytest-cov==2.5.1
coverage==4.5.1

celery[redis]==4.2.1
numpy==1.14.1
pandas==0.22.0
cython==0.27.3
scipy==1.0.0
scikit-learn==0.19.1
